chore: move the browser-only modules under src/web
rewind-ui, disc-visualiser and config join the other page modules; only
import paths change.
